pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bnel-mef3-server"
version = "0.0.2"
description = "A gRPC server and client for efficient, concurrent access to MEF3 files."
authors = [
    { name = "BNEL Team", email = "your@email.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy~=2.3.3",
    "grpcio~=1.75.0",
    "protobuf~=6.32.1",
    "grpcio-tools~=1.75.0",
    "mef_tools==1.2.3",
]

[project.optional-dependencies]
test = [
    "pytest~=8.4.2",
    "pytest-order",
]
docs = [
    "sphinx",
    "sphinxcontrib-napoleon",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools]
include-package-data = true

